+++ EPS Device Extensions.indigoPlugin/Contents/Server Plugin/lib/eps.py
@@ -5,6 +5,7 @@
 
 import indigo
 import logging
+import sys
 
 import ext
 
@@ -15,7 +16,7 @@ from support import support
 
 
 class eps:
-	VERSION = "2.4.2"
+	VERSION = "2.4.4"
 	
 	#
 	# Initialize the  class
@@ -24,6 +25,7 @@ class eps:
 		if plugin is None: return # pre-loading before init
 		
 		try:
+			#self.memory_summary ()
 			self.plugin = plugin
 		
 			# Sets the log format in the plugin and across all modules that can reference the plugin
@@ -134,6 +136,19 @@ class eps:
 		if ext.valueValid (plugin.pluginPrefs, "debugMode"): 
 			self.logger.info(u"Upgraded plugin preferences from pre-Indigo 7, depreciated preferences removed")
 			del plugin.pluginPrefs["debugMode"]
+			
+	#
+	# Only for use in dev environment with Pympler installed so we can track memory usage
+	def memory_summary(self):
+		# Only import Pympler when we need it. We don't want it to
+		# affect our process if we never call memory_summary.
+		
+		caller = sys._getframe(1).f_code.co_name # So we can reference the caller
+		
+		from pympler import summary, muppy
+		mem_summary = summary.summarize(muppy.get_objects())
+		rows = summary.format_(mem_summary)
+		indigo.server.log ('\n\nCALLED BY: ' + caller + '\n\n' + '\n'.join(rows)	)	
 				
 		
 	#